Improve the CI (#38)
* Do not push or pull translations on pull_request event

Allow to run tests without wasting API requests

* Install tx directly into PATH directory

Avoid a untracked downloaded README in the VCS tree

* Remove check for CI variable in update.sh

* Debug update.sh

* Rename locales dirname on-the-fly

* Revert "Rename locales dirname on-the-fly"

This reverts commit 568a2dd68dd98378cd7f641abf5fcd3c0be7f0c4.

This issue is already solved by another commit, hence reverting.

* Revert "Debug update.sh"

This reverts commit 1647b7bc4a89cd3abcd06a8839eba62dec4fdedd.

* Move languages up in update.sh

* Simplify main.yml

This also drops forcing master. On push or schedule,
it will run on master as per event trigger settings.
On pull request, it should run in the branch of the
pull request.

* Use sed instead of bash substitution

GHA complained ./locale/update.sh: 8: Bad substitution

* Fix command execution

* Do not fail when there is nothing to commit

* Remove skip ci from main.yml

I don't see a particular reason for this, and it could
prevent other workflows that we could want to be ran.

* Use git push instead of an action

* Add workflow_dispatch for manual run

* Create test-translations.yml

Centralized build and lint for all languages

* Add status badges to README.rst

* Fix reST syntax for badges in README.rst

* Add tx config to .gitignore

The current workflow is already to not commit it.
With this entry we save a line in update.sh

* update.sh: move POT and config into generate_template.sh

This step is required for later deploying a script that
removes obsolete resources in Transifex

* Revert .tx/config in .gitignore

It did not work for some reason

* Add lock-translations.py and check-obsoletes CI job

* Fix path for cat

* Install tx tool in check-obsoletes job

* lock-translations.py: add dry-run for pull requests

* Actually remove POT and resources from update job

* Be verbose on the commit

Useful for pull requests

* Add transifex-python and python-dateutil to requirements.txt

* Fix git rm command

* Add -j auto to sphinx-build

* Also remove obsolete PO files

The update.sh deletes PO files before pulling new translations. Then we have modified files (updated) and deleted (obsolete, nothing pulled). This commit removes from git the deleted files.

* Remove unneeded re module import

* apt update on graphviz install, TZ=UTC on make html

* Set python version to 3 instead of 3.9 in main.yml

* Improve CI trigger event name in update.sh

Replace [[ with [ for POSIX shell compatibility,
and check for empty variable (e.g. in own local shell)